chore: Rename protocol/io.ts to protocol/command.ts
