+++ protocol/command.ts
